c.sccharts: Support for non-compound loop bodies. Support for for loops.
For loops are now better supported than before, where now the approach
to show the text for the initializer, condition, and update in the for
loop are shown in text. This is now done by looking up the text directly
in the source file, so that the back translation cannot go wrong
anymore. Also added support for initializations in the initializer
clause of for loops, as they are now added as declarations (without
their initialization) to its region.