c.sccharts: connect all in- and outputs of functions correctly.
Some parameters may be modified by a function call (such as arrays or
pointers). With this change, all arrays modified by a known function or
potentially modified by an unknown function are taken through the
function call and put as an output of the function again for further
sequential references to it.
Unified function call behavior.
Fixed buggy view when calling a function twice in the same function.