c.sccharts: added support for nested return statements
Any return statement contained within control loops such as if, switch,
while, etc. is now supported. In each state, the returned value is
assigned to a 'res' value. If for some state there are multiple return
values (for example because multiple ifs may return), a new state
merging all returns into a final return value for the state is created.