scg: Fixed missing prio statement in Java Prio code generation.
This fixes the ABRO-vars test.
This fix only handles a single case, the underlying problem seems to be
a structural issue in the code generation process. Prio statement should
not be added before a node that rises the priority but when leaving to a
node with higher priority as the fix does!